requirements.txt
streamlit
openai
pandas
pypdf
python-dotenv
